Quick and dirty support for H5Part IO
Adds binary HDF5 file writing using H5Part format
use cmake -DSPH_EXA_WITH_H5PART=ON to enable it and make sure
that HDF5 can be found by cmake on the system.

This enables some code that was already present and contributed
to the MPIFileUtils section. I have backported it to an older
H5Part version that can be included as a small lib (extern/h5part).

The IO as implemented makes a copy of the arrays prior to writing
which is not usually efficient and should be looked at.

Not tested in parallel yet, but should not require significant changes
to ensure that it works from multiple ranks.