Update add_m1.py
The original script was adding the fields to the last step, which are sorted alphabetically, not numerically, in the HDF5, which led to the problem that the fields were added to wrong step.
There were also warnings regarding attempts to use a numpy array in a context that requires a scalar. Specifically with minDt. This is solved, too.
Finally, the mass of the particles is also added. It is calculated directly as 1./num_particles, which is correct for some scenarios, but not all.